requirements.txt
boto3
paramiko
pyyaml
awscli
